Refactor return messages for server. Finish unit tests for server.
